// API_Client/client/models/create-attribute-option-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Shop API
 * REST API for the online shop backend
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface CreateAttributeOptionDto {
    /**
     * ოფციის მნიშვნელობა ქართულად
     */
    'valueKa': string;
    /**
     * ოფციის მნიშვნელობა ინგლისურად
     */
    'valueEn': string;
    /**
     * უნიკალური კოდი (attribute-ის ფარგლებში), მხოლოდ ლათინური ასოები/ციფრები/ტირე
     */
    'code': string;
    /**
     * დალაგების რიგი
     */
    'sortOrder'?: number;
}